Create cleanliness_check.py
using OpenCV to capture an image of the object and then using TensorFlow to classify the object as clean or not clean

OpenCV is used to capture an image from the camera (index 0) and store it in the frame variable. The image is then resized to 32x32 to match the input shape required by the model. Finally, the predict_cleanliness function is called to classify the object as clean or not clean based on the image.